+++ src/bittytax/conv/parsers/__init__.py
@@ -54,6 +54,7 @@ from . import (
     ledgerlive,
     liquid,
     mercatox,
+    mymonero,
     nault,
     nexo,
     okx,