+++ CHANGELOG.md
@@ -90,6 +90,7 @@
 - Ledger Live parser: new data file format added.
 - Exodus parser: added "deposit (failed)" type.
 - Conversion tool: added new parser for Eternl wallet.
+- Conversion tool: added new parser for MyMonero wallet.
 ### Changed
 - Conversion tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))
 - Accounting tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))